Fix all GMloader scripts with new standards from https://portmaster.games/packaging.html Added [ -f "/etc/os-release" ] && source "/etc/os-release" Added new logging method "exec > >(tee "$GAMEDIR/log.txt") 2>&1" Encapsulate Pipewirefix for Jelos in IF statements Fix false LD_LIBRARY_PATH that exluded JELOS Added file check for  data.win & game.win mv Changed Names of Chip n Dale, thecurseofissyos and others

+++ ports/apocrunner/Apoc Runner.sh
@@ -19,23 +19,24 @@ printf "\033c" > /dev/tty0
 printf "\033c" > /dev/tty1
 
 GAMEDIR="/$directory/ports/apocrunner"
-if [ -f "/etc/os-release" ]; then
-  source "/etc/os-release"
-fi 
 
-if [ "$OS_NAME" != "JELOS" ]; then
-  export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/apocrunner/libs"
-fi
- 
+[ -f "/etc/os-release" ] && source "/etc/os-release"
+
 cd "$GAMEDIR"
-export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
-export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+
+exec > >(tee "$GAMEDIR/log.txt") 2>&1
+
+if [ "$OS_NAME" == "JELOS" ]; then
+  export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
+  export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+fi
+
 export GMLOADER_DEPTH_DISABLE=1
 export GMLOADER_SAVEDIR="$GAMEDIR/gamedata/"
+export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/apocrunner/libs"
 
-mv gamedata/data.win gamedata/game.droid
-mv gamedata/game.win gamedata/game.droid
-
+[ -f "./gamedata/data.win" ] && mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/game.win" ] && mv gamedata/game.win gamedata/game.droid
 
 $ESUDO chmod 666 /dev/uinput
 $GPTOKEYB "gmloader" textinput &
@@ -43,8 +44,7 @@ echo "Loading, please wait... " > /dev/tty0
 
 $ESUDO chmod +x "$GAMEDIR/gmloader"
 
-
-./gmloader apocrunner.apk |& tee log.txt /dev/tty0
+./gmloader apocrunner.apk
 
 $ESUDO kill -9 "$(pidof gptokeyb)"
 $ESUDO systemctl restart oga_events &
