+++ ports/thecurseofissyos/The Curse of Issyos.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-# PORTMASTER: zeroranger.zip, Zeroranger.sh
 if [ -d "/opt/system/Tools/PortMaster/" ]; then
   controlfolder="/opt/system/Tools/PortMaster"
 elif [ -d "/opt/tools/PortMaster/" ]; then
@@ -9,6 +8,7 @@ else
 fi
 
 source "$controlfolder/control.txt"
+[ -f "/etc/os-release" ] && source "/etc/os-release"
 
 get_controls
 
@@ -17,31 +17,30 @@ $ESUDO chmod 666 /dev/tty1
 printf "\033c" > /dev/tty0
 printf "\033c" > /dev/tty1
 
-GAMEDIR="/$directory/ports/CHIPNDALE"
-if [ -f "/etc/os-release" ]; then
-  source "/etc/os-release"
-fi 
-
-if [ "$OS_NAME" != "JELOS" ]; then
-  export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/CHIPNDALE/lib"
-fi
+GAMEDIR="/$directory/ports/thecurseofissyos"
+exec > >(tee "$GAMEDIR/log.txt") 2>&1
  
 cd "$GAMEDIR"
 
 export GMLOADER_DEPTH_DISABLE=1
 export GMLOADER_SAVEDIR="$GAMEDIR/gamedata/"
-export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
-export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/thecurseofissyos/lib"
 
-mv gamedata/data.win gamedata/game.droid
+if [ "$OS_NAME" == "JELOS" ]; then
+  export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
+  export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+fi
+
+[ -f "./gamedata/data.win" ] && mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/game.win" ] && mv gamedata/game.win gamedata/game.droid
 
 $ESUDO chmod 666 /dev/uinput
-$GPTOKEYB "gmloader" -c "chip.gptk" &
+$GPTOKEYB "gmloader" &
 echo "Loading, please wait... " > /dev/tty0
 
 $ESUDO chmod +x "$GAMEDIR/gmloader"
 
-./gmloader chip.apk |& tee log.txt /dev/tty0
+./gmloader IssyosWrapper.apk
 
 $ESUDO kill -9 "$(pidof gptokeyb)"
 $ESUDO systemctl restart oga_events &
