+++ ports/zeroranger/Zeroranger.sh
@@ -8,6 +8,7 @@ else
 fi
 
 source "$controlfolder/control.txt"
+[ -f "/etc/os-release" ] && source "/etc/os-release"
 
 get_controls
 
@@ -17,13 +18,8 @@ printf "\033c" > /dev/tty0
 printf "\033c" > /dev/tty1
 
 GAMEDIR="/$directory/ports/zeroranger"
-if [ -f "/etc/os-release" ]; then
-  source "/etc/os-release"
-fi 
 
-if [ "$OS_NAME" != "JELOS" ]; then
-  export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/zeroranger/lib"
-fi
+exec > >(tee "$GAMEDIR/log.txt") 2>&1
  
 cd "$GAMEDIR"
 
@@ -34,14 +30,17 @@ if [ ! -f flagfile ]; then
    touch flagfile
 fi
 
+if [ "$OS_NAME" == "JELOS" ]; then
+  export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
+  export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+fi
 
 export GMLOADER_DEPTH_DISABLE=1
 export GMLOADER_SAVEDIR="$GAMEDIR/gamedata/"
-export SPA_PLUGIN_DIR="/usr/lib32/spa-0.2"
-export PIPEWIRE_MODULE_DIR="/usr/lib32/pipewire-0.3/"
+export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$directory/ports/zeroranger/lib"
 
-
-mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/data.win" ] && mv gamedata/data.win gamedata/game.droid
+[ -f "./gamedata/game.win" ] && mv gamedata/game.win gamedata/game.droid
 
 $ESUDO chmod 666 /dev/uinput
 $GPTOKEYB "gmloader" &
@@ -49,8 +48,7 @@ echo "Loading, please wait... " > /dev/tty0
 
 $ESUDO chmod +x "$GAMEDIR/gmloader"
 
-
-./gmloader zeroranger.apk |& tee log.txt /dev/tty0
+./gmloader zeroranger.apk
 
 $ESUDO kill -9 "$(pidof gptokeyb)"
 $ESUDO systemctl restart oga_events &
